# backend/openapi.yml
openapi: 3.1.0
info:
  title: Video Station API
  version: 1.0.0
  description: VOD 스트리밍 서비스 API
servers:
  - url: http://localhost:8080
    description: 로컬 개발 서버
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    RegisterRequest:
      type: object
      required: [email, password, name, nickname]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
        nickname:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          description: Access Token 만료 시간 (초)
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        name:
          type: string
        nickname:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, VIEWER]
        status:
          type: string
          enum: [ACTIVE, INACTIVE, BANNED]
        profileImageUrl:
          type: string
        createdAt:
          type: string
          format: date-time

    # === 동영상 ===
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [인증]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '409':
          description: 이메일 중복
  /api/v1/auth/login:
    post:
      tags: [인증]
      summary: 로그인
      description: Access Token은 응답 body, Refresh Token은 HttpOnly 쿠키로 발급
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          headers:
            Set-Cookie:
              description: Refresh Token (HttpOnly)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 인증 실패
  /api/v1/auth/refresh:
    post:
      tags: [인증]
      summary: 토큰 갱신
      description: 쿠키의 Refresh Token으로 새 Access Token 발급
      responses:
        '200':
          description: 갱신 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: Refresh Token 만료 또는 없음
  /api/v1/auth/me:
    get:
      tags: [인증]
      summary: 현재 사용자 정보
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

  # === 사용자 관리 API (SUPER_ADMIN) ===
...