# backend/src/main/java/com/videostation/application/dto/LoginRequest.java
package com.videostation.application.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {}